컴포넌트 폴더 구조 개선
- 도메인, 공용 폴더 구분
