+++ app/(auth)/register/page.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import RegisterForm from "@/components/forms/register-form";
+import RegisterForm from "@/components/domains/login/register-form";
 
 export const metadata = {
   title: "회원가입",
