+++ app/(tabs)/posts/loading.tsx
@@ -1,4 +1,4 @@
-import PostPreviewSkeleton from "@/components/skeletons/post-preview";
+import PostPreviewSkeleton from "@/components/domains/post/preview-skeleton";
 
 export default function PostsLoading() {
   return [...Array(10)].map((_, index) => <PostPreviewSkeleton key={index} />);
