+++ app/posts/[id]/layout.tsx
@@ -2,7 +2,7 @@ import Header from "@/components/layouts/header";
 import db from "@/libs/db";
 import { notFound } from "next/navigation";
 import getSession from "@/libs/session";
-import PostDetailMoreBtns from "@/components/buttons/post-detail-more-buttons";
+import PostDetailMoreBtns from "@/components/domains/post/more-buttons";
 
 interface LayoutProps {
   children: React.ReactNode;
