+++ app/users/[id]/friends/page.tsx
@@ -1,5 +1,5 @@
 import EmptyState from "@/components/layouts/empty-state";
-import ProfileImage from "@/components/profile-image";
+import ProfileImage from "@/components/domains/profile/image";
 import db from "@/libs/db";
 import getSession from "@/libs/session";
 import { ChevronRightIcon } from "@heroicons/react/24/solid";
