+++ app/users/[id]/page.tsx
@@ -1,5 +1,5 @@
 import UserInfo from "@/components/layouts/user-info";
-import UserTimeline from "@/components/user-timeline";
+import UserTimeline from "@/components/domains/post/user-timeline";
 import db from "@/libs/db";
 import getSession from "@/libs/session";
 import { notFound } from "next/navigation";
