+++ components/domains/post/preview-skeleton.tsx
